requirements.txt
Jinja2-3.1.5 
MarkupSafe-3.0.2 
Werkzeug-3.1.3 
aniso8601-9.0.1 
blinker-1.9.0 
click-8.1.8 
flask-3.1.0 
flask-cors-5.0.0 
flask-restful-0.3.10 
itsdangerous-2.2.0 
pytz-2024.2 
six-1.17.0
flask-sqlalchemy-3.1.1 
sqlalchemy-2.0.36
flask-login-0.6.3